[ADD] : added the recent version of OpenXR 1.0.31

+++ 3rdparty/OpenXR_SDK/.azure-pipelines/nuget/stage_nuget.ps1
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020 The Khronos Group Inc.
+# Copyright (c) 2019-2023, The Khronos Group Inc.
 # SPDX-License-Identifier: Apache-2.0
 param(
     [Parameter(Mandatory = $true, HelpMessage = "Path to unzipped openxr_loader_windows OpenXR-SDK release asset")]
@@ -55,15 +55,17 @@ Copy-Item (Join-Path $SDKRelease "include") (Join-Path $NugetStaging "include") 
 #
 function CopyLoader($Platform) {
     $PlatformSDKPath = Join-Path $SDKRelease "$Platform"
-    $NuGetPlatformPath = Join-Path $NugetStaging "native/$Platform/release"
+    if (Test-Path $PlatformSDKPath) {
+        $NuGetPlatformPath = Join-Path $NugetStaging "native/$Platform/release"
 
-    $NugetLibPath = Join-Path $NuGetPlatformPath "lib"
-    New-Item $NugetLibPath -ItemType "directory" -Force
-    Copy-Item (Join-Path $PlatformSDKPath "lib/openxr_loader.lib") $NugetLibPath
+        $NugetLibPath = Join-Path $NuGetPlatformPath "lib"
+        New-Item $NugetLibPath -ItemType "directory" -Force
+        Copy-Item (Join-Path $PlatformSDKPath "lib/openxr_loader.lib") $NugetLibPath
 
-    $NugetBinPath = Join-Path $NuGetPlatformPath "bin"
-    New-Item $NugetBinPath -ItemType "directory" -Force
-    Copy-Item (Join-Path $PlatformSDKPath "bin/openxr_loader.dll") $NugetBinPath
+        $NugetBinPath = Join-Path $NuGetPlatformPath "bin"
+        New-Item $NugetBinPath -ItemType "directory" -Force
+        Copy-Item (Join-Path $PlatformSDKPath "bin/openxr_loader.dll") $NugetBinPath
+    }
 }
 
 # Currently there are no non-UWP ARM/ARM64 binaries available from the SDK release.
